+++ apps/api/v2/src/modules/api-key/api-key.service.ts
