+++ apps/api/v2/src/modules/auth/auth-api-key.strategy.ts
