+++ apps/api/v2/src/modules/auth/auth.module.ts
