+++ apps/api/v2/src/modules/booking/dtos/create-booking.ts
