+++ apps/api/v2/test/app.e2e-spec.ts
@@ -3,7 +3,7 @@ import type { TestingModule } from "@nestjs/testing";
 import { Test } from "@nestjs/testing";
 import * as request from "supertest";
 
-import { AppModule } from "./../src/app.module";
+import { AppModule } from "../src/app.module";
 
 describe("AppController (e2e)", () => {
   let app: INestApplication;
